[#13] geth: import and unlock HDWallet keys
- Add python script to derive the keys from the mnemonic.
- Load keys into geth in bash wrapper script.
- Remove keystore directory when shutting down geth.
- Share TEST_MNENOMIC between hardhat and geth.

The geth wrapper bash script `run-geth` does not look very maintainable
at the moment but we currently also don't know if doing the signing
inside geth is all that useful for us. The rust code does the signing on
the client side at the moment (it does not rely on the geth node to
sign). So I think we can postpone finding a more elegant solution.

Closes #13.