[#37] [WIP] Call blake2 pre-compile
This is currently not functional, feel free to discard.

The approach taken was to flip the
endiannes of the uint types so that when abi.encodePacked is called on
them they end up in little endian. This may not be the right approach.

However some of the functions or parts thereof could be useful.