Use zerok_lib with keccak256 nullifier hash
Compiles, hash tests fail because solidity is still using blake2.

The smart contract needs to be simplified because we now have a 32-byte
hash function output size (and therefore tree height).

If this is not desirable and we would like to try with 64-byte / 512
tree height first we could somehow stretch the output of the keccak hash
function instead.

These tests currently fail

failures:
    nullifiers_merkle_tree::tests::test_blake2b_branch
    nullifiers_merkle_tree::tests::test_blake2b_elem
    nullifiers_merkle_tree::tests::test_blake2b_leaf
    nullifiers_merkle_tree::tests::test_terminal_node_value_empty