Make the test `test_terminal_node_value_non_empty`.
Problem: the function NullifiersMerkleTree.to_bool_array does not match the rust function `into_bits`.
The difference is that for each byte (chunk of 8 bits) the order of the bits are reversed.

Solution: reverse the order of the bits.

Also the test was changed in order to call directly `terminal_node.value()`.
In order to do this it was required to make the function SetMerkleTerminalNode::value() public.
See https://gitlab.com/translucence/systems/system/-/commit/b53f5d83a01724f9e54043a9d96d592f65296f94