NullifiersMerkleTree check set (non)membership
Happy path of set membership seems to be working

The commit also contains a draft plan for how update the tree with new
nullifiers. This is a basic unoptimized version that only works with one
nullifier at a time.

The idea of this draft outline is to

1. supply the new nullifier and a new root
2. ensure the nullifier is not in the existing set
3. ensure the nullifier is in the new set/root
4. store the new root

The rust test code currently contains a lot of code duplication put
basic tests for set membership are passing also for non-empty trees.