[#129] Do not reject block if nullifiers spent
- Add a boolean mask marking which notes should be included in the
  block.
- Process nullifiers of transactions in original order of transactions.

Don't like the choice of mask arrays but given the lack of in memory
hashmap or growing array this is the best thing I could think of so far
today.

Depending on how the data will be encoded for the plonk verification we
may want to drop the masks later and instead encode the data directly as
we loop over the notes.