Check transfer note does not have burn prefix
Added an extra function even though it is very small. I think it makes
the main validation loop more readable to not litter it with require
statemens. We can add all the checks (for merkle root and anything else
we would like to add) into the _check[NoteType] functions.

Close #137.