Avoid hitting 24kB contract size limit
Deploy RescueLib as standalone libary

- Link RescueLib if placeholder in bytecode

Use `async_recursion` crate to be able to call async deploy function
recursively. This is suggested by the compiler upon failing to compile
because of the recursion. Since we only need it in one place and
currently only have one non-inlined library (RescueLib) we could find
ways around it. For example we could inline the deploy code for the
recue library where it's used.

- Re-enable previously disabled hardhat tests.
- Re-enable 1M optimizer runs.
- Document how to connect to linked library via env var.