Allow extra calldata for block submission
Adds a function that accepts an additional bytes argument which just
calls the block submitting functions.

The block submitting entity can use the extra data to store anything in
calldata. Our current use case is to store `ReceiverMemo`s and their
signature.

Includes an example test to show how to submit a block with receiver
memos and fetch and decode the memos after block submission.

When this feature is not/no-longer useful the existing function
`submitCapeBlock` can be used.

Close #360