Separate `await_transaction` from `transfer_token`. (#627)
This allows the caller to choose how they want to wait on the
TransactionReceipt. In the CapeBackend tests, we wait for _all_
wallets to process the transaction and require it to succeed. In
the random wallet tests, we just wait for the sender to process
the transaction, and we don't panic if it fails.