Remove temporary directory after tests
Avoid calling `TempDir::into_path`, otherwise the automatic removal when
the variable goes out of scope is disabled.

The wallet tests still don't clean up because this would require some
refactoring, for example passing in the address book temporary directory
to `create_test_network` so that the temporary directory could be
cleaned up when it goes out of scope. We could also consider having an
in-memory implementation so that tests outside the address_book could
use that instead.