Support named keystores with an endpoint to list them.
The bulk if this change is reworking the way storage is configured,
so that there is now one parameter for the root storage directory
for all wallets. This replaces `~/.espresso/cape/wallet` if given.
This directory is then used to store both the last used wallet path
(`$CAPE_WALLET_STORAGE/last_wallet_path`) and named keystores
(`$CAPE_WALLET_STORAGE/keystores/`).

We then extend the `newwallet` and `openwallet` endpoints to take a
:name instead of a :path, which automatically stores the created
wallet in `$CAPE_WALLET_STORAGE/keystores/$NAME`. We add an endpoint
`listkeystores` to list all of the keystore names in the keystores
directory.