Fix web server tests
I moved the web server related files out of the lib target and into
the web_server bin target, so that #[cfg(test)] can be used to select
between the mock backend and the real backend. This makes it harder to
accidentally select the mock backend then using feature flags. I had
dreams of making the whole web server generic on the backend type, but
alas, this triggers an ICE.

Also moved the web server bin from wallet/src/bin to wallet/src/web_server,
since module files cannot be placed in bin/ or accessed from a binary if
they are placed in src/.