Stopping geth with SIGHUP does not persist state
Use SIGTERM to stop geth.

Why was `SIGHUP` used in the first place?

We use SIGHUP to kill the `run-geth` script in our test scripts. The
reason being that this is the only thing I found that works reliably for
those scripts. In the `cape-demo-docker` script we are invoking geth
directly and `SIGTERM` is the right signal to send to geth to get a
clean shutdown that persists the state.