Distribute multiple records to faucet servers in faucet-shower. (#935)
This allows a faucet server to have multiple transfers in flight at
the same time, since not all of its balance is locked up in a single
record. We can do this perfectly, because the faucet server always
transfers records of the same size with 0 fees, so there is no risk
of fragmentation.

This is not the best solution to this problem: better would be to
have the faucet service itself check, on startup, if it has any
records larger than 2*CAPE_FAUCET_GRANT_SIZE, and break them up if
so. However, there is currently no wallet API to spend specific
records, so this could get tricky. Doing this in faucet-shower when
setting up the faucet should be good enough as a stop-gap until we
can implement the better solution.