Do not add sponsored asset to library until we know it is ok.
Previously, the buildsponsor endpoint (and the corresponding
`build_sponsor` wallet method) would add the new AssetInfo to the
wallet's library before returning the `sol::AssetDefinition` to be
submitted to the contract. This can lead to confusing UX if the user
cancels the transaction before the asset is actually sponsored, or
if the Ethereum transaction is rejected or reverted, because it
would lead to an asset in the asset library which doesn't actually
correspond to anything.

Now, `buildsponsor` just returns the serialized AssetInfo instead of
adding it to the wallet. The asset can be added to the wallet at any
time after the sponsor transaction succeeds by calling `importasset`.