Include pending transaction in nonce count
This allows the relayer to include more than one Ethereum txn per block.

However it does not fix all race condition and submitting a new
transaction before the previous one is in the mempool will still lead to
an error.

We could use ethers-rs' nonce manager middleware to perform the nonce
count locally but this also present a some new challenges like handling
the case where the count is out of sync.