Problem: relayer Eth txn easily runs out of gas
Solution: specify the gas limit manually.

Adds a `CAPE_RELAYER_GAS_LIMIT` env var so we can adjust the value
without changing code if needed.

It would be better to create a custom type for the gas limit to avoid
confusing it with other u64s. It may also we worth it to group all the
Ethereum config together instead of passing around so many variables.
In light of time pressure I went with what was quicker to implement.