Run multiple instances of random wallet script
This works. There seem to be issues with getting assets from the faucet
if the random wallet instances are started too quickly after each other.

It could be further combined with tools to simulate network delays and
packet drops and longer ethereum block times (for example by setting
GETH_PERIOD=15 in `cape-demo-local` for 15 seconds block time).