Emit block data fields individually
If we emit the CapeBlock struct directly this leads to a compilation
error. The workaround to emit ABI encoded bytes instead fails to decode
the CapeBlock in rust struct at runtime. This workaround emits the
ABI encoded fields of CapeBlock individually.

The function `decode_cape_block` uses an ugly workaround to deal with
the problem of not being able to import `BlockCommittedFilter` because
that type exists both for `CAPE` and `TestCAPE` and is therefore
ambiguous.