Verify Plonk verifier and merkle tree on etherscan
This is expected to fail on the current revision of this repo because of
some non-functional but bytecode affecting changes to the merkle tree
contract in #1109 that have not been deployed.

To verify one can checkout 0d13dec7b86a6afaaa511228924ee5092bd05fc6 and
run the `etherscan-verify` of this commit but I already ran it and
all contracts are now verified.

To test the script the easiest way is to do a throwaway deployment on
Goerli and run the `etherscan-verify` script.

Close #1123