Persist number of grants given to each key.
This fixes a race condition where we could be shut down in the middle
of transferring several grants to a key. After restarting, previously,
we would start granting to this key from scratch, potentially granting
more than we intended and even, if we get shut down frequently, not
making progress at all.

Note, there is still a small race condition: if we get shut down just
after building a transfer but before incrementing the count of transfers
to that key (a very small window of time) we may repeat the grant the
next time we are restarted. This is a minor problem, as it means at
worst tranfserring an extra grant per restart. To fix this, we need
to synchronize between the faucet's storage and the wallet library
storage, which I think requires some kind of composabilities between
AtomicStores.