Ensure non-deployer is not deployer
If `CAPE_WEB3_PROVIDER_URL` is `get_funded_client()` returns the same
client each time its called.

Use a `call` instead of a `send` to avoid requiring a funded client.

This still ensures that the call fails for the right reason (the
function was called from the wrong account) because we match against the
revert reason.