relayer: add extra gas to transaction
We previously set the limit manually but this is tricky because gas cost
is different on arbitrum. By adding the 10M to the estimate instead it
should work for both L1's and arbitrum.

Remove the test that tests if the extra gas has effect. The test would
be a bit involved to rewrite right now because we can't set a too low
gas limit directly anymore.